+++ example/routes/chip/index.tsx
@@ -1,26 +1,36 @@
-// import { Chip, Header, Main, Text } from '../../mod.ts';
-// import XIcon from 'icons/x.tsx';
+import {
+  Card,
+  Chip,
+  Header,
+  Main,
+  Separator,
+  Text,
+} from '../../../components.ts';
+import { ChipSelect } from '../../../islands.ts';
 
-export default function TestText() {
+export default function ChipTest() {
   return (
     <>
-      {
-        /* <Header layout_type='center'>
+      <Header layout_type='focus'>
         <Text type='title'>Chip tests</Text>
       </Header>
-      <Main layout_type='center'>
-        <Chip icon={<XIcon size={12} />} />
-        <Chip
-          icon={<XIcon size={12} />}
-          onRemove={() => console.log('x')}
-        />
-        <Chip
-          icon={<XIcon size={12} />}
-          onRemove={() => console.log('x')}
-          content='Lorem ipsum dolor, sit amet consectetur adipisicing elit. Rerum ut accusamus tempore, dignissimos quaerat voluptatem odio necessitatibus reiciendis deleniti corporis non numquam vero eum autem nam sapiente. Deserunt, eius molestias?'
-        />
-      </Main> */
-      }
+      <Main layout_type='focus'>
+        <Card>
+          <Text type='subheading' noMargins>Chip</Text>
+          <div class='flex gap-4'>
+            <Chip content='Test Chip' />
+            <Chip content='Selected Chip' selected />
+          </div>
+          <Separator />
+          <Text type='subheading' noMargins>ChiplistSelect</Text>
+          <ChipSelect
+            initialChips={[
+              { content: 'test chip' },
+              { content: 'test 2 chip' },
+            ]}
+          />
+        </Card>
+      </Main>
     </>
   );
 }